requirements.txt
playwright==1.42.0
pandas==2.2.1
openpyxl==3.1.2
greenlet==3.0.3 # Explicitly add greenlet
python-telegram-bot==22.7
psycopg2-binary
